feat: default port when exposing deployment (#41)
Currently when we expose a deployment via a service using `deployment.expose`, we require a port the service will bind to. 

This PR makes the port optional, defaulting to the port of the first container in the deployment, similarly to how we configure the `targetPort` of the service. Funnily enough, this is a breaking change since we move from a positional to a keyword argument.

BREAKING CHANGE: `deployment.expose` and `service.addDeployment` now accept a `port` as part of the options, and not a positional argument.