feat: expose service via ingress (#51)
I was looking at our canonical [example](https://cdk8s.io/docs/latest/plus/ingress/) for using `Ingress` with CDK8s+ and felt we could simplify this even more. 

This PR introduces an `exposeViaIngress` method on a **service**, which will create the necessary `IngressBackend` and the corresponding rule on the `Ingress`. 

In addition, add an `exposeViaIngress` method on a **deployment**, which will first expose it via a service, and then expose the service via an `Ingress`. 

Following this PR, our example can be simplified to:

```ts
new kplus.Deployment(this, text, {
  containers: [
    {
      image: 'hashicorp/http-echo',
      args: [ '-text', 'hello ingress' ],
      port: 5678
    }
  ]
}).exposeViaIngress('/hello');
```

BREAKING CHANGE: `deployment.expose` renamed to `deployment.exposeViaService`