chore: hardcode k8s spec version (#242)
Every branch in this project corresponds to a different k8s version. This meant that every branch needs to set a different value for the `SPEC_VERSION` constant in the `.projenrc.js` file.

We tried to avoid differences in this file between branches by being smart and auto detecting the version based on the branch name. But this turned out to be more of a hassle then a benefit since it required a naming convention for feature branches.

> See https://github.com/cdk8s-team/cdk8s-plus/blob/k8s-22/main/CONTRIBUTING.md#making-a-pull-request

Also, our current mechanism doesn't capture tag names. Which means that checking out to a tag would always built the default version, instead of the one the tag was created from.

Instead of this madness, lets just hardcode the version in the file and thats it. The diff it creates between `.projenrc.js` files across branches is very minimal and isn't likely to create conflicts when performing backports.

Signed-off-by: Christopher Rybicki <rybickic@amazon.com>