feat: security context (#455)
`SecurityContext` for pods and containers. I included only the properties that I thought were most common and that I could quickly understand. If we see the need to add additional ones, we'll do that at a later time. 

The default values correspond to the values applied by Kubernetes if the property is not specified.

Fixes https://github.com/cdk8s-team/cdk8s-plus/issues/348