chore: automatic backport (#498)
Add a workflow for automatically backporting changes to our various release branches.

The way it works is for every PR that needs backporting, add a `backport <branch-name>` label to the PR.
Once the PR is closed and merged, the workflow will create backport PRs for the branch names specified in the labels.

Currently we put this burden on repository maintainers, i.e external contributors don't have to worry about and should operate as normal.

Signed-off-by: iliapolo <epolon@amazon.com>