chore: fix backport workflow (#655)
There's currently somewhat of a mess with the backport workflow, which doesn't actually work. 

The action we use is [tibdex/backport@v1](https://github.com/tibdex/backport), which I mistakenly thought was internally using the [backport](https://github.com/sqren/backport) tool i've been testing this out with. The correct action to use is actually [sqren/backport-github-action](https://github.com/sqren/backport-github-action). 

However, the action doesn't currently use the correct version of `backport` that contains the [fix](https://github.com/sqren/backport/pull/336) we need for DCO to pass.

I asked the author to cut a new release, but he hasn't responded yet. I then realized the action is not really necessary, since most of the logic is implemented in the `backport` tool (finally a well moduled action). So this PR configures the workflow to use the `backport` tool as a normal `run` step, via projen.  


Signed-off-by: Eli Polonsky <epolon@amazon.com>