chore: clean backport  (#695)
When the backport task runs inside a cloned repository directory, it doesn't do a fresh clone, and relies on the local git status to be up to date. 

Instead of trying to update it, always run the backport task in an empty directory. This causes some performance hit, but is worth the stability.