chore: clean backport  (#695)
When the backport task runs inside a cloned repository directory, it doesn't do a fresh clone, and relies on the local git status to be up to date.

Instead of trying to update it, always run the backport task in an empty directory. This causes some performance hit, but is worth the stability.

(cherry picked from commit 3be2efcbbf0f5ca1771089ddbb6d7260d8cffe5c)
Signed-off-by: Eli Polonsky <epolon@amazon.com>