feat: everything is a construct (#747)
Currently, imported resources are not treated as constructs. We didn't initially implement them this way because there was no need, and it makes the API slightly more awkward.

Lately, we've been adding more and more of those, and a use case for them being constructs has risen, when those resources are being used for constructing actual managed resources.

This PR makes it so all imported resources (and selected as well) are constructs, and thus have stable id's we can use.
Even though this hurts API ergonomics a little bit, its worth it for the simplicity and flexibility it gives us. This is also in-line with how the AWS CDK treats imported resources, and in general, having to pass scopes and id's to classes in a construct library is a well known practice.

BREAKING CHANGE: You must now pass a `scope` and `id` to imported and selected resources

(cherry picked from commit 24d9f6117fa8673a38e2bb6686986f14f001361c)
Signed-off-by: Eli Polonsky <epolon@amazon.com>