feat(cronjob): introduce `CronJob` construct (#1052)
This PR adds the Kubernetes [CronJob Resource L2](https://kubernetes.io/docs/reference/kubernetes-api/workload-resources/cron-job-v1/) to cdk8s-plus. This is part of the effort to make cdk8s+ GA. 

Having this L2 abstraction would enable the customer to provision Kubernetes CronJob Resource using cdk8s+ and that would generate/add it to the Kubernetes manifest.

Resolves https://github.com/cdk8s-team/cdk8s-plus/issues/376
Resolves https://github.com/cdk8s-team/cdk8s-plus/issues/813
Resolves https://github.com/cdk8s-team/cdk8s-plus/issues/814

## Next Steps

Since timezone is behind a feature gate and works with K8s-24, we would need to fix the backport PR.

## DCO

Signed-off-by: Vinayak Kukreja <vinakuk@amazon.com>