feat(container): default container startup probe (#1093)
Startup probe help Kubernetes determine if the container is ready and no other probe is initiated till this check is complete. Currently we leave it as undefined in the manifest if the values are not provided.
Now, if a port number is provided by the user, that will be knocked on to understand status of the container.

Resolves https://github.com/cdk8s-team/cdk8s-plus/issues/808

Signed-off-by: Vinayak Kukreja <vinakuk@amazon.com>
(cherry picked from commit 839b14a5b90707962586b04cf259cb8988b7212e)
Signed-off-by: Vinayak Kukreja <78971045+vinayak-kukreja@users.noreply.github.com>

# Conflicts:
#	test/container.test.ts