fix(hpa): unable to target a deployment with a container that has volume mounts (#1438)
The `_toKube` method of `HorizontalPodAutoscaler` was populating the result with a reference to the `.containers` property of a deployment. This created a resolving loop since `containers` can contain constructs, which refer contain circular references to each other via the `host` and `defaultChild` properties. 

Remove the `.containers` from the result of `_toKube`, it was oversight, and is not necessary. In general as a rule of thumb, `_toKube` methods should always result in a JSON serializable object.

Fixes https://github.com/cdk8s-team/cdk8s-plus/issues/1423