feat: Add volumes from Secrets Store
K8s has a CSI driver for generic 'secrets store' data. The CRD for that
defines a `SecretProviderClass` resource which has a parameters block to
pass arbitrary data to specific providers. Once a `SecretProviderClass`
object has been created, it can be mounted on a container as a volume.

Of specific interest to CDK8s is the AWS provider which allows secrets
to be automatically loaded from AWS Secrets Manager or SSM Parameter
store, including pulling specific keys out from JSON data, but all of
that logic is encapsulated in the `SecretProviderClass`. As such, this
modification should work with _any_ provider, not just the AWS one.

Signed-off-by: Dan Meyers <d.meyers@infinitefunspace.co.uk>