chore: decouple optional and required properties for a container (#1894)
This simple PR decouples the optional properties of a container from the required ones.

Rationale:

We use this library in combination with CDK's capabilities to handle compiling, bundling and uploading images to ECR. In which case we expect our users to provide all other properties of a container but the `image` property which is abstracted away.