fix(statefulsets): port name not included in headless service (#3077)
If you have two different containers in a Stateful Set that both require port access Kubernetes (v.1.27) will fail with the following error:

```bash
The Service "SERVICE_NAME" is invalid:
* spec.ports[0].name: Required value
* spec.ports[1].name: Required value
```

This is because during creation of the headless service the name of the ports is not passed. This will also happen if you have a container with multiple ports.