fix(pod): allow sidecar containers to have probes (#4483)
https://kubernetes.io/docs/concepts/workloads/pods/init-containers/#differences-from-regular-containers

if we check docs, sidecar containers support readiness, liveness and startup probes. I added an extra check to make sure the init container is not sidecar for probe checks.

this way it will be possible to add probes for initcontainers when they are sidecar.

edit: this should be backported to k8s-29/main as well

Fixes #4338