fix: revert jsii-docgen to produce smaller docs (#4594)
In https://github.com/cdk8s-team/cdk8s-plus/pull/4577 `jsii-docgen` got updated to the current v10 version line.

Unfortunately this update brings the new addition of summary tables for parameters and class methods to generated documentation. For this package that means an increase in file-size from 650kb per language to ca. 4.5mb per language. This increase in turn breaks the generation of the `cdk8s` website because `mkdocs` cannot handle such large files.

As an immediate solution, this PR reverts the version of `jsii-docgen` back to v3.

Mid-term, the way forward would probably be to use the `--split-by-submodules` options and adjust the `cdk8s` website script to accommodate for it. Or if that doesn't work (because submodule docs are still to large), we will need to add new features into `jsii-docgen` to allow for more control of the output. Finally, the docs process here (or in `cdk8s`) could switch to the json output format and implement their own, custom markdown rendering.

(cherry picked from commit 23453ae77c7946c0613e29591c6280c921f5572d)
Signed-off-by: Momo Kornher <mail@moritzkornher.de>